+++ src/sll/api/_generated_raw.c
@@ -654,6 +654,31 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_file_copy_raw(sll_object
 
 
 
+__SLL_API_TYPE_sll_api_file_delete sll_api_file_delete(__SLL_API_ARGS_sll_api_file_delete);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_file_delete_raw(sll_object_t*const* al,sll_arg_count_t all){
+	sll_object_t* a=NULL;
+	if (all>0){
+		a=*(al+0);
+		if (SLL_OBJECT_GET_TYPE(a)==SLL_OBJECT_TYPE_STRING){
+			SLL_ACQUIRE(a);
+		}
+		else{
+			a=sll_operator_cast(a,sll_static_int[SLL_OBJECT_TYPE_STRING]);
+		}
+	}
+	else{
+		a=SLL_CREATE();
+		a->t=SLL_OBJECT_TYPE_STRING;
+		sll_string_create(0,&(a->dt.s));
+	}
+	sll_bool_t out=sll_api_file_delete(&(a->dt.s));
+	SLL_RELEASE(a);
+	SLL_ACQUIRE(sll_static_int[out]);
+	return sll_static_int[out];
+}
+
+
+
 __SLL_API_TYPE_sll_api_file_flush sll_api_file_flush(__SLL_API_ARGS_sll_api_file_flush);
 __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_file_flush_raw(sll_object_t*const* al,sll_arg_count_t all){
 	sll_object_t* a=NULL;
@@ -5077,6 +5102,10 @@ static const internal_function_t _ifunc_data_ptr[]={
 		sll_api_file_copy_raw
 	},
 	{
+		SLL_CHAR("sll:file_delete"),
+		sll_api_file_delete_raw
+	},
+	{
 		SLL_CHAR("sll:file_flush"),
 		sll_api_file_flush_raw
 	},
@@ -5636,5 +5665,5 @@ static const internal_function_t _ifunc_data_ptr[]={
 
 
 
-const sll_function_index_t _ifunc_size=158;
+const sll_function_index_t _ifunc_size=159;
 const internal_function_t* _ifunc_data=(const internal_function_t*)(&_ifunc_data_ptr);